Add RISC-V architecture support with SBI, memory management, platform initialization, and syscall definitions
- Implement SBI calls for console I/O, timer management, and system shutdown in `lib.rs`.
- Create memory management traits for user and kernel address spaces in `memory.rs`.
- Initialize Virt platform devices in `platform.rs`.
- Define RISC-V syscall numbers in `syscall.rs` following Linux conventions.
