--- os/src/arch/loongarch/lib/mod.rs
@@ -1,3 +1,0 @@
-//! LoongArch64 库模块
-
-pub mod platform;
