--- os/src/arch/loongarch/memory/mod.rs
@@ -1,6 +1,0 @@
-//! LoongArch 内存子系统 VirtualMemory trait 实现
-//!
-//! 包含 `LoongArch64KernelAddressSpace` 和 `LoongArch64ProcessAddressSpace`。
-
-pub mod address_space;
-pub mod mmu;
