--- os/src/arch/loongarch/syscall/mod.rs
@@ -1,3 +1,0 @@
-//! LoongArch64 系统调用号定义
-mod syscall_number;
-pub use syscall_number::*;
