--- os/src/arch/riscv/lib/mod.rs
@@ -1,1 +1,0 @@
-pub mod sbi;
