--- os/src/arch/riscv/memory/mod.rs
@@ -1,6 +1,0 @@
-//! RISC-V 内存子系统 VirtualMemory trait 实现
-//!
-//! 包含 `Riscv64KernelAddressSpace` 和 `Riscv64ProcessAddressSpace`。
-
-pub mod address_space;
-pub mod mmu;
