--- os/src/arch/riscv/platform/mod.rs
@@ -1,4 +1,0 @@
-//! 平台相关模块
-pub mod virt;
-
-pub use virt::init;
