--- os/src/arch/riscv/syscall/mod.rs
@@ -1,3 +1,0 @@
-//! RISC-V 系统调用号定义
-mod syscall_number;
-pub use syscall_number::*;
