+++ os/src/mm/memory_space/space.rs
@@ -1,7 +1,7 @@
 use core::cmp::Ordering;
 
 use crate::arch::mm::{paddr_to_vaddr, vaddr_to_paddr};
-use crate::arch::platform::virt::MEMORY_END;
+use crate::arch::platform::MEMORY_END;
 use crate::config::{
     MAX_USER_HEAP_SIZE, PAGE_SIZE, USER_SIGRETURN_TRAMPOLINE, USER_STACK_SIZE,
     USER_STACK_TOP,
