+++ os/src/mm/mod.rs
@@ -22,7 +22,7 @@ pub use frame_allocator::init_frame_allocator;
 pub use global_allocator::init_heap;
 
 use crate::arch::ArchImpl;
-use crate::arch::platform::virt::MEMORY_END;
+use crate::arch::platform::MEMORY_END;
 use crate::config::PAGE_SIZE;
 use crate::earlyprintln;
 use crate::mm::address::{Ppn, UsizeConvert};
